patch

+++ quick-start.sh
@@ -34,7 +34,7 @@ zsh "$HOME/dotfiles/hammerspoon/dock-switching/dock-switcher.sh" --load home
 
 #-------------------------------------------------------------------------------
 # CREATE SYMLINKS
-DOTFILE_FOLDER="$(dirname "$0")"
+DOTFILE_FOLDER="$HOME/.config/"
 
 # zsh
 [[ -e ~/.zshrc ]] && rm -rf ~/.zshrc
@@ -73,20 +73,12 @@ ln -sf "$DOTFILE_FOLDER/warp" ~/.warp
 # INFO: Marta as default folder opener set in Duti Script
 ln -sf /Applications/Marta.app/Contents/Resources/launcher /opt/homebrew/bin/marta
 MARTA_DIR=~"/Library/Application Support/org.yanex.marta"
-if [[ -e "$MARTA_DIR" ]]; then
-	rm -rf "$MARTA_DIR"
-else
-	mkdir -p "$MARTA_DIR"
-fi
+[[ -e "$MARTA_DIR" ]] && rm -rf "$MARTA_DIR"
 ln -sf "$DOTFILE_FOLDER/Marta" "$MARTA_DIR"
 
 # Espanso
 ESPANSO_DIR=~"/Library/Application Support/espanso"
-if [[ -e "$ESPANSO_DIR" ]]; then
-	rm -rf "$ESPANSO_DIR"
-else
-	mkdir -p "$ESPANSO_DIR"
-fi
+[[ -e "$ESPANSO_DIR" ]] && rm -rf "$ESPANSO_DIR"
 ln -sf "$DOTFILE_FOLDER/espanso/" "$ESPANSO_DIR"
 
 #-------------------------------------------------------------------------------
@@ -96,11 +88,7 @@ ln -sf "$DOTFILE_FOLDER/espanso/" "$ESPANSO_DIR"
 
 # Brave PWAs
 # BROWSER="Brave Browser"
-# if [[ -e ~"/Applications/$BROWSER Apps.localized" ]] ; then
-# 	rm -rf ~"/Applications/$BROWSER Apps.localized"
-# else
-# 	mkdir -p ~"/Applications/$BROWSER Apps.localized"
-# fi
+# [[ -e ~"/Applications/$BROWSER Apps.localized" ]] && rm -rf ~"/Applications/$BROWSER Apps.localized"
 # ln -sf ~"/Library/Mobile Documents/com~apple~CloudDocs/Dotfolder/$BROWSER Apps.localized/" ~"/Applications/$BROWSER Apps.localized"
 
 # to keep private stuff out of the dotfile repo
